Update INSTALL.md
I think combining OpenAI, ROS and Gazebo is a brilliant idea. I hope more people can use this repository. The installation is always the first barrier between the users and the framework. I met some problems during the installation, and I notice that most of them have been discussed in issues like #36 and #38, but those solutions are not in the installation instruction. I tested them on Ubuntu 16.04, ROS kinetic, and update the installation instruction of Ubuntu 16.04. But I am not sure if other versions have the same problem.

ros-kinetic-ar-track-alvar-msgs is from https://github.com/erlerobot/gym-gazebo/issues/36. Not sure if this happens on Ubuntu 18.04, ROS melodic.

The others exist in the installation instruction for Ubuntu 18.04, but they are missed in installation instruction for Ubuntu 16.04.